Fix pixel aspect ratio from AVFoundation.

+++ MP42/MP42AVFImporter.m
@@ -135,7 +135,7 @@
                     CFDictionaryRef pixelAspectRatioFromCMFormatDescription = CMFormatDescriptionGetExtension(formatDescription, kCMFormatDescriptionExtension_PixelAspectRatio);
 
                     if (pixelAspectRatioFromCMFormatDescription) {
-                        NSInteger hSpacing, vSpacing;
+                        int hSpacing, vSpacing;
                         CFNumberGetValue(CFDictionaryGetValue(pixelAspectRatioFromCMFormatDescription, kCMFormatDescriptionKey_PixelAspectRatioHorizontalSpacing), kCFNumberIntType, &hSpacing);
                         CFNumberGetValue(CFDictionaryGetValue(pixelAspectRatioFromCMFormatDescription, kCMFormatDescriptionKey_PixelAspectRatioVerticalSpacing), kCFNumberIntType, &vSpacing);
                         videoTrack.hSpacing = hSpacing;
